Add c2a cards to landing page header
