+++ website/src/css/custom.css
@@ -49,7 +49,7 @@
   color: var(--ifm-color-primary);
   padding-right: 0.6rem;
   padding-left: 0.8rem;
-  padding-bottom: 0.2rem;
+  padding-bottom: 0.5rem;
   font-weight: 600;
   font-size: 1.4rem;
 }
