+++ website/src/pages/index.module.css
@@ -1,10 +1,3 @@
-/* stylelint-disable docusaurus/copyright-header */
-
-/**
- * CSS files with the .module.css suffix will be treated as CSS modules
- * and scoped locally.
- */
-
 .heroBanner {
   padding: 4rem 0;
   text-align: center;
@@ -18,9 +11,10 @@
   }
 }
 
-.buttons {
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.cardHeader {
+  color: var(--ifm-color-gray-900)
 }
 
+.cardBody {
+  color: var(--ifm-color-gray-700)
+}
